requirements.txt
streamlit==1.31.0
pandas==2.2.0
requests==2.31.0
anthropic==0.18.1
fpdf2==2.7.9
openpyxl==3.1.2
